Renamed package to conversationitem and placed relevant classes there
