+++ layer-atlas/src/main/java/com/layer/ui/ConversationsRecyclerView.java
@@ -28,14 +28,14 @@ import android.view.View;
 import com.layer.sdk.LayerClient;
 import com.layer.sdk.messaging.Conversation;
 import com.layer.ui.adapters.ConversationsAdapter;
-import com.layer.ui.conversation.OnConversationItemClickListener;
+import com.layer.ui.conversationitem.OnConversationItemClickListener;
 import com.layer.ui.messagetypes.CellFactory;
 import com.layer.ui.messagetypes.location.LocationCellFactory;
 import com.layer.ui.messagetypes.singlepartimage.SinglePartImageCellFactory;
 import com.layer.ui.messagetypes.text.TextCellFactory;
 import com.layer.ui.messagetypes.threepartimage.ThreePartImageCellFactory;
 import com.layer.ui.util.AvatarStyle;
-import com.layer.ui.util.ConversationItemFormatter;
+import com.layer.ui.conversationitem.ConversationItemFormatter;
 import com.layer.ui.util.ConversationStyle;
 import com.layer.ui.util.itemanimators.NoChangeAnimator;
 import com.layer.ui.util.views.SwipeableItem;
