+++ layer-atlas/src/main/java/com/layer/ui/adapters/ConversationsAdapter.java
@@ -14,11 +14,11 @@ import com.layer.sdk.query.Predicate;
 import com.layer.sdk.query.Query;
 import com.layer.sdk.query.RecyclerViewController;
 import com.layer.sdk.query.SortDescriptor;
-import com.layer.ui.conversation.ConversationItemViewModel;
-import com.layer.ui.conversation.OnConversationItemClickListener;
+import com.layer.ui.conversationitem.ConversationItemViewModel;
+import com.layer.ui.conversationitem.OnConversationItemClickListener;
 import com.layer.ui.databinding.UiConversationItemBinding;
 import com.layer.ui.messagetypes.CellFactory;
-import com.layer.ui.util.ConversationItemFormatter;
+import com.layer.ui.conversationitem.ConversationItemFormatter;
 import com.layer.ui.util.ConversationStyle;
 import com.layer.ui.util.IdentityRecyclerViewEventListener;
 import com.layer.ui.util.Log;
