+++ layer-atlas/src/main/java/com/layer/ui/conversationitem/OnConversationItemClickListener.java
@@ -1,4 +1,4 @@
-package com.layer.ui.conversation;
+package com.layer.ui.conversationitem;
 
 import com.layer.sdk.messaging.Conversation;
